[Bugfix]主要解决SCLog里float与CGFloat混用导致的bug

+++ SCBasicComponents/Classes/Tools/SCLog/SCLogManager.m
@@ -97,7 +97,7 @@ NSString * const SCLogDebugTag = @"[DEBUG]";
     fclose(stderr);
 }
 
-- (void)setMaxLogFileSize:(CGFloat)maxLogFileSize
+- (void)setMaxLogFileSize:(float)maxLogFileSize
 {
     _maxLogFileSize = maxLogFileSize;
     if (maxLogFileSize > 0) {
